+++ src/js/moss.js
@@ -1,6 +1,6 @@
 /*!
  * moss-ui - The user interface framework for front-end.
- * @version v0.1.4
+ * @version v0.1.5
  * @link https://github.com/archco/moss-ui
  * @license MIT
  */