+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.1.5] - 2017-07-29
 ### Fixed
 - Fixed problem in rendering navbar. #20
 - Fixed helper functions. #21
@@ -41,7 +43,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - [README.md](README.md)
 - [docs/README.md](docs/README.md)
 
-[Unreleased]: https://github.com/archco/moss-ui/compare/v0.1.4...HEAD
+[Unreleased]: https://github.com/archco/moss-ui/compare/v0.1.5...HEAD
+[0.1.5]: https://github.com/archco/moss-ui/compare/v0.1.4...v0.1.5
 [0.1.4]: https://github.com/archco/moss-ui/compare/v0.1.3...v0.1.4
 [0.1.3]: https://github.com/archco/moss-ui/compare/v0.1.2...v0.1.3
 [0.1.2]: https://github.com/archco/moss-ui/compare/v0.1.1...v0.1.2